requirements.txt
# Core
fastapi>=0.132.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Image Processing
Pillow>=10.4.0
opencv-python>=4.10.0
numpy>=1.26.0

# Deep Learning
torch>=2.5.0
torchvision>=0.20.0
transformers>=4.45.0
# accelerate — indirect dep (BiRefNet). CVE-2025-14925: LOCAL-only, no fix available.
# Risk is low: only trusted HuggingFace checkpoints are loaded.
accelerate>=0.34.0

# Super Resolution
super-image>=0.1.7

# Alpha Matting
pymatting>=1.1.10

# Testing (CVE-2025-71176: UNIX-only, not applicable on Windows)
pytest>=8.3.0
pytest-asyncio>=0.24.0
